requirements.txt
flask>=3.0
gunicorn
python-dotenv
sqlalchemy>=2.0
flask-sqlalchemy
flask-migrate
psycopg2-binary
pgvector
alembic
celery[redis]
redis
psutil
openai-whisper
# Using standard transformers + torch
torch
transformers
sentence-transformers
openai
anthropic
tiktoken
pymupdf
yt-dlp
pydub
mcp
pydantic-settings
python-multipart
requests
langchain-text-splitters
EbookLib
beautifulsoup4
docker>=7.0.0
ddgs
markdown2
xhtml2pdf
